Feat: develop merge를 위한 reply 기능 부분 Admin 활성화

+++ backend/src/main/java/com/pofo/backend/domain/reply/service/ReplyService.java
@@ -5,7 +5,6 @@ import com.pofo.backend.domain.inquiry.repository.InquiryRepository;
 import com.pofo.backend.domain.reply.dto.request.ReplyCreateRequest;
 import com.pofo.backend.domain.reply.dto.request.ReplyUpdateRequest;
 import com.pofo.backend.domain.reply.dto.response.ReplyCreateResponse;
-import com.pofo.backend.domain.reply.dto.response.ReplyDeleteResponse;
 import com.pofo.backend.domain.reply.dto.response.ReplyDetailResponse;
 import com.pofo.backend.domain.reply.dto.response.ReplyUpdateResponse;
 import com.pofo.backend.domain.reply.entity.Reply;
@@ -23,11 +22,11 @@ public class ReplyService {
     private final InquiryRepository inquiryRepository;
 
     @Transactional
-    public ReplyCreateResponse create(Long id, ReplyCreateRequest replyCreateRequest) {
+    public ReplyCreateResponse create(Long id, ReplyCreateRequest replyCreateRequest, Admin admin) {
 
-//        if (admin == null) {
-//            throw new ReplyException("관리자 정보가 유효하지 않습니다.");
-//        }
+        if (admin == null) {
+            throw new ReplyException("관리자 정보가 유효하지 않습니다.");
+        }
 
         Inquiry inquiry = this.inquiryRepository.findById(id)
                 .orElseThrow(() -> new ReplyException("문의사항을 찾을 수 없습니다."));
@@ -38,7 +37,7 @@ public class ReplyService {
 
         try {
             Reply reply = Reply.builder()
-//                    .admin(admin)
+                    .admin(admin)
                     .inquiry(inquiry)
                     .content(replyCreateRequest.getContent())
                     .build();
@@ -52,10 +51,14 @@ public class ReplyService {
     }
 
     @Transactional
-    public ReplyUpdateResponse update(Long inquiryId, Long replyId, ReplyUpdateRequest replyUpdateRequest) {
+    public ReplyUpdateResponse update(Long inquiryId, Long replyId, ReplyUpdateRequest replyUpdateRequest, Admin admin) {
 
         Reply reply = this.replyRepository.findByInquiryIdAndId(inquiryId, replyId)
                 .orElseThrow(() -> new ReplyException("해당 답변을 찾을 수 없습니다."));
+
+        if (!reply.getAdmin().equals(admin)) {
+            throw new UnauthorizedActionException("답변을 수정할 권한이 없습니다.");
+        }
 
         try {
             reply.update(replyUpdateRequest.getContent());
@@ -66,19 +69,20 @@ public class ReplyService {
     }
 
     @Transactional
-    public ReplyDeleteResponse delete(Long inquiryId, Long replyId) {
+    public void delete(Long inquiryId, Long replyId, Admin admin) {
 
         Reply reply = this.replyRepository.findByInquiryIdAndId(inquiryId, replyId)
                 .orElseThrow(() -> new ReplyException("해당 답변을 찾을 수 없습니다."));
 
+        if (!reply.getAdmin().equals(admin)) {
+            throw new UnauthorizedActionException("답변을 수정할 권한이 없습니다.");
+        }
+
         try {
             this.replyRepository.delete(reply);
-
             Inquiry inquiry = this.inquiryRepository.findById(inquiryId)
                     .orElseThrow(() -> new ReplyException("문의사항을 찾을 수 없습니다."));
             inquiry.updateResponse(false);
-
-            return new ReplyDeleteResponse();
         } catch (Exception e) {
             throw new ReplyException("답변 삭제 중 오류가 발생했습니다. 원인: " + e.getMessage());
         }
